app/requirements.txt
python-telegram-bot~=20.0a2
redvid
